requirements.txt
Flask==1.0.2
gunicorn==19.6.0
Werkzeug<0.13.0,>=0.12.0
GoogleAppEngineCloudStorageClient==1.9.22.1
google-cloud-storage==1.8.0